Extract shared code for cross-platform reuse
Move platform-agnostic business logic (types, GitHub API, utils, hooks)
into a shared/ directory that both the web app and upcoming iOS app can
import from. Web app files become thin re-export wrappers so all
existing imports continue to work unchanged.

Introduce StorageAdapter interface to abstract localStorage for mobile
compatibility (AsyncStorage/SecureStore). Web app uses webStorage
adapter that wraps localStorage with Promise-based API.

Refs #101
