+++ shared/storage.ts
@@ -1,0 +1,6 @@
+/** Platform-agnostic storage interface. Web uses localStorage, mobile uses AsyncStorage/SecureStore. */
+export interface StorageAdapter {
+  getItem(key: string): Promise<string | null>;
+  setItem(key: string, value: string): Promise<void>;
+  removeItem(key: string): Promise<void>;
+}
